Add boilerplate for CAPH test automation.
This change is progress towards creating test automation
where CAPH can be tested running against a mock version
of cloud agent.

Changes include:

1. Modify the Makefile and co. to allow publishing container image to
custom container registry.

2. Create scripts for setting up k3d cluster with CAPI/CAPH. These will
be used by the tests but they are also handy for manual testing
(especially if you want to avoid Kind to avoid the requirement to disable
the swap file on your dev box).

3. Create helper library/package for creating a mock cloud agent gRPC
server.

4. Create a mockcloudagent executable that reports an error for all requests.
This can be useful for manual testing. Though in the future, it could be
extended with more interesting logic.

5. Create the boilerplate for the tests, which simply sets up a mock cloud agent
and a CAPI/CAPH deployment (using k3d) and ensures CAPH can communicate with
the mock cloud agent.